+++ cs336_data/__init__.py
@@ -1,3 +1,0 @@
-import importlib.metadata
-
-__version__ = importlib.metadata.version("cs336-data")